add 'po_to_mo' script in locales >X1X2
